improve audit

+++ judge-api/src/main/resources/db/migration/V2__create_attachment_table.sql
@@ -8,7 +8,7 @@ create table attachment (
 );
 
 create table attachment_aud (
-    rev int not null,
+    rev bigint not null,
     revtype smallint not null,
     id uuid not null,
     created_at timestamp not null,
